+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## [`master`](https://github.com/elastic/eui/tree/master)
 
+- Added a new `EuiWindowEvent` component for declarative, safe management of `window` event listeners
+- Changed `Flyout` component to close on ESC keypress even if the flyout does not have focus
+
+## [`master`](https://github.com/elastic/eui/tree/master)
+
 - Added `zIndexAdjustment` to `EuiPopover` which allows tweaking the popover content's `z-index` ([#1097](https://github.com/elastic/eui/pull/1097))
 - Added new `EuiSuperSelect` component and `hasArrow` prop to `EuiPopover` ([#921](https://github.com/elastic/eui/pull/921))